Added file type check that only allows mp3 audio files to be uploaded as recordings.

+++ wordpress/wp-content/plugins/gc-custom-post-types/recording-post-type/meta.php
@@ -86,6 +86,13 @@ function update_recording_file( $post_id ) {
         return $post_id;
     } else {
 
+        $supported_file_types = array('audio/mpeg');
+        $uploaded_file_type = wp_check_filetype(basename($_FILES['recording-file']['name']))['type'];
+
+        if ( ! in_array( $uploaded_file_type, $supported_file_types )) {
+            return $post_id;
+        }
+
         if ( $meta_value !== "" ) {
             unlink( $meta_value['file'] );
         }
